udp_reader listening and printing
udp_reader listens on port 2368 and 8308 and prints out the data
received. It also hosts the sockets on the local ip address (127.0.0.1
will not work.) In order to look up the local ip address, we run
ifconfig and grep for the line we need. This is bad, but was easy.