SimpleHDLViewer writes our own compressed version of pcd
The viewer now writes data to a file which can be processed later by a
script. The name of each output file is the time (in milliseconds)
from the top of the hour. The data format is 3 floats (representing x,
y, and z position) followed by a short (representing intensity)