gps_hdl_grabber now parses position packets
This patch allows gps_hdl_grabber to parse packets coming from port
8308 in pcap files. When we get a new packet we use the length to
determine which port it is coming from. If the length is 512 it came
from port 8308 and is a positioning packet. In order to parse these
packets, I added a 'tag' to the first byte of data. If this tag is an
'l' the packet is a laser packet, if it is a 'p' the packet is a
positioning packet.

We use positioning packets to name frames. Frames are formatted in the
same way, but the name is now in microseconds from epoch time. This is
accomplished by combining information from the position data packet on
8308. This packet contains an nmea sentence which we parse to get the
date and hour of day. We then combine that information with the time
stamp from each laser sweep.